Move variable collectors into their own folder.
